fix: import Raycast 2.0.5 exports
Refs abue-ammar/tinycast#321
